chore(ecmascript): Colocate the rest of the heap impls
