+++ nova_vm/src/ecmascript/builtins/date/data.rs
@@ -1,4 +1,7 @@
-use crate::ecmascript::types::OrdinaryObject;
+use crate::{
+    ecmascript::types::OrdinaryObject,
+    heap::{CompactionLists, HeapMarkAndSweep, WorkQueues},
+};
 use std::time::SystemTime;
 
 #[derive(Debug, Clone, Copy)]
@@ -13,5 +16,15 @@ impl DateHeapData {
             object_index: None,
             date: None,
         }
+    }
+}
+
+impl HeapMarkAndSweep for DateHeapData {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        self.object_index.mark_values(queues);
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        self.object_index.sweep_values(compactions);
     }
 }
