+++ nova_vm/src/ecmascript/builtins/finalization_registry/data.rs
@@ -1,6 +1,19 @@
-use crate::ecmascript::types::OrdinaryObject;
+use crate::{
+    ecmascript::types::OrdinaryObject,
+    heap::{CompactionLists, HeapMarkAndSweep, WorkQueues},
+};
 
 #[derive(Debug, Clone, Default)]
 pub struct FinalizationRegistryHeapData {
     pub(crate) object_index: Option<OrdinaryObject>,
+}
+
+impl HeapMarkAndSweep for FinalizationRegistryHeapData {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        self.object_index.mark_values(queues);
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        self.object_index.sweep_values(compactions);
+    }
 }
