+++ nova_vm/src/ecmascript/builtins/set/data.rs
@@ -1,4 +1,7 @@
-use crate::ecmascript::types::{OrdinaryObject, Value};
+use crate::{
+    ecmascript::types::{OrdinaryObject, Value},
+    heap::{CompactionLists, HeapMarkAndSweep, WorkQueues},
+};
 
 #[derive(Debug, Clone, Default)]
 pub struct SetHeapData {
@@ -15,4 +18,18 @@ pub struct SetHeapData {
     // TODO: When an non-terminal (start or end) iterator exists for the Set,
     // the items in the map cannot be compacted.
     // pub(crate) observed: bool;
+}
+
+impl HeapMarkAndSweep for SetHeapData {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        self.object_index.mark_values(queues);
+        self.set.iter().for_each(|value| value.mark_values(queues));
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        self.object_index.sweep_values(compactions);
+        self.set
+            .iter_mut()
+            .for_each(|value| value.sweep_values(compactions));
+    }
 }
