+++ nova_vm/src/ecmascript/builtins/weak_map/data.rs
@@ -1,4 +1,7 @@
-use crate::ecmascript::types::{OrdinaryObject, Value};
+use crate::{
+    ecmascript::types::{OrdinaryObject, Value},
+    heap::{CompactionLists, HeapMarkAndSweep, WorkQueues},
+};
 
 #[derive(Debug, Clone, Default)]
 pub struct WeakMapHeapData {
@@ -16,4 +19,26 @@ pub struct WeakMapHeapData {
     // TODO: When an non-terminal (start or end) iterator exists for the Map,
     // the items in the map cannot be compacted.
     // pub(crate) observed: bool;
+}
+
+impl HeapMarkAndSweep for WeakMapHeapData {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        self.object_index.mark_values(queues);
+        for ele in &self.keys {
+            ele.mark_values(queues);
+        }
+        for ele in &self.values {
+            ele.mark_values(queues);
+        }
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        self.object_index.sweep_values(compactions);
+        for ele in &mut self.keys {
+            ele.sweep_values(compactions);
+        }
+        for ele in &mut self.values {
+            ele.sweep_values(compactions);
+        }
+    }
 }
