+++ nova_vm/src/ecmascript/builtins/weak_set.rs
@@ -8,7 +8,10 @@ use crate::{
             OrdinaryObjectInternalSlots, PropertyDescriptor, PropertyKey, Value,
         },
     },
-    heap::{indexes::WeakSetIndex, ObjectEntry, ObjectEntryPropertyDescriptor},
+    heap::{
+        indexes::{BaseIndex, WeakSetIndex},
+        CompactionLists, HeapMarkAndSweep, ObjectEntry, ObjectEntryPropertyDescriptor, WorkQueues,
+    },
     Heap,
 };
 
@@ -18,8 +21,19 @@ use super::ordinary::ordinary_set_prototype_of_check_loop;
 
 pub mod data;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[repr(transparent)]
 pub struct WeakSet(pub(crate) WeakSetIndex);
+
+impl WeakSet {
+    pub(crate) const fn _def() -> Self {
+        Self(BaseIndex::from_u32_index(0))
+    }
+
+    pub(crate) const fn get_index(self) -> usize {
+        self.0.into_index()
+    }
+}
 
 impl From<WeakSet> for WeakSetIndex {
     fn from(val: WeakSet) -> Self {
@@ -47,13 +61,13 @@ impl IntoObject for WeakSet {
 
 impl From<WeakSet> for Value {
     fn from(val: WeakSet) -> Self {
-        Value::WeakSet(val.0)
+        Value::WeakSet(val)
     }
 }
 
 impl From<WeakSet> for Object {
     fn from(val: WeakSet) -> Self {
-        Object::WeakSet(val.0)
+        Object::WeakSet(val)
     }
 }
 
@@ -274,5 +288,15 @@ impl InternalMethods for WeakSet {
         } else {
             Ok(vec![])
         }
+    }
+}
+
+impl HeapMarkAndSweep for WeakSetHeapData {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        self.object_index.mark_values(queues);
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        self.object_index.sweep_values(compactions);
     }
 }
