+++ nova_vm/src/ecmascript/execution.rs
@@ -8,9 +8,8 @@ pub use agent::{Agent, JsResult};
 pub use default_host_hooks::DefaultHostHooks;
 pub(crate) use environments::{
     get_this_environment, new_declarative_environment, new_function_environment,
-    DeclarativeEnvironment, DeclarativeEnvironmentIndex, EnvironmentIndex, Environments,
-    FunctionEnvironment, FunctionEnvironmentIndex, GlobalEnvironment, GlobalEnvironmentIndex,
-    ModuleEnvironmentIndex, ObjectEnvironment, ObjectEnvironmentIndex, PrivateEnvironment,
+    DeclarativeEnvironmentIndex, EnvironmentIndex, Environments, FunctionEnvironmentIndex,
+    GlobalEnvironment, GlobalEnvironmentIndex, ModuleEnvironmentIndex, ObjectEnvironmentIndex,
     PrivateEnvironmentIndex, ThisBindingStatus,
 };
 pub(crate) use execution_context::*;
@@ -18,4 +17,4 @@ pub use realm::{
     create_realm, initialize_default_realm, initialize_host_defined_realm, set_realm_global_object,
     Realm,
 };
-pub(crate) use realm::{Intrinsics, ProtoIntrinsics, RealmIdentifier};
+pub(crate) use realm::{ProtoIntrinsics, RealmIdentifier};
