+++ nova_vm/src/ecmascript/execution/environments/private_environment.rs
@@ -1,4 +1,7 @@
-use crate::ecmascript::types::{Function, Value};
+use crate::{
+    ecmascript::types::{Function, Value},
+    heap::{CompactionLists, HeapMarkAndSweep, WorkQueues},
+};
 use std::collections::HashMap;
 
 use super::PrivateEnvironmentIndex;
@@ -41,6 +44,16 @@ pub struct PrivateEnvironment {
     names: HashMap<String, PrivateName>,
 }
 
+impl HeapMarkAndSweep for PrivateEnvironment {
+    fn mark_values(&self, _queues: &mut WorkQueues) {
+        todo!()
+    }
+
+    fn sweep_values(&mut self, _compactions: &CompactionLists) {
+        todo!()
+    }
+}
+
 /// ### [9.2.1.1 NewPrivateEnvironment ( outerPrivEnv )](https://tc39.es/ecma262/#sec-newprivateenvironment)
 ///
 /// The abstract operation NewPrivateEnvironment takes argument outerPrivEnv (a
@@ -57,4 +70,14 @@ pub(crate) fn new_private_environment(
         names: Default::default(),
     }
     // }.
+}
+
+impl HeapMarkAndSweep for PrivateEnvironmentIndex {
+    fn mark_values(&self, _queues: &mut WorkQueues) {
+        todo!()
+    }
+
+    fn sweep_values(&mut self, _compactions: &CompactionLists) {
+        todo!()
+    }
 }
