+++ nova_vm/src/ecmascript/scripts_and_modules.rs
@@ -1,4 +1,8 @@
-use self::{module::ModuleIdentifier, script::ScriptIdentifier};
+use crate::heap::{CompactionLists, HeapMarkAndSweep, WorkQueues};
+
+use self::script::ScriptIdentifier;
+
+use super::builtins::module::Module;
 
 pub mod module;
 pub mod script;
@@ -6,5 +10,21 @@ pub mod script;
 #[derive(Debug, Clone, Copy)]
 pub(crate) enum ScriptOrModule {
     Script(ScriptIdentifier),
-    Module(ModuleIdentifier),
+    Module(Module),
+}
+
+impl HeapMarkAndSweep for ScriptOrModule {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        match self {
+            ScriptOrModule::Script(idx) => idx.mark_values(queues),
+            ScriptOrModule::Module(idx) => idx.mark_values(queues),
+        }
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        match self {
+            ScriptOrModule::Script(idx) => idx.sweep_values(compactions),
+            ScriptOrModule::Module(idx) => idx.sweep_values(compactions),
+        }
+    }
 }
