+++ nova_vm/src/ecmascript/types/language/object.rs
@@ -32,20 +32,19 @@ use super::{
 use crate::{
     ecmascript::{
         builtins::{
-            bound_function::BoundFunction, data_view::DataView, date::Date, error::Error, map::Map,
-            primitive_objects::PrimitiveObject, set::Set, shared_array_buffer::SharedArrayBuffer,
-            ArgumentsList, Array, ArrayBuffer, BuiltinFunction, ECMAScriptFunction,
+            bound_function::BoundFunction, data_view::DataView, date::Date,
+            embedder_object::EmbedderObject, error::Error,
+            finalization_registry::FinalizationRegistry, map::Map, module::Module,
+            primitive_objects::PrimitiveObject, promise::Promise, proxy::Proxy, regexp::RegExp,
+            set::Set, shared_array_buffer::SharedArrayBuffer, weak_map::WeakMap, weak_ref::WeakRef,
+            weak_set::WeakSet, ArgumentsList, Array, ArrayBuffer, BuiltinFunction,
+            ECMAScriptFunction,
         },
         execution::{Agent, JsResult},
-        scripts_and_modules::module::ModuleIdentifier,
         types::PropertyDescriptor,
     },
     heap::{
-        indexes::{
-            ArrayIndex, EmbedderObjectIndex, FinalizationRegistryIndex, MapIndex, ObjectIndex,
-            PromiseIndex, ProxyIndex, RegExpIndex, SetIndex, TypedArrayIndex, WeakMapIndex,
-            WeakRefIndex, WeakSetIndex,
-        },
+        indexes::{ArrayIndex, ObjectIndex, TypedArrayIndex},
         CompactionLists, HeapMarkAndSweep, WorkQueues,
     },
 };
@@ -84,16 +83,16 @@ pub enum Object {
     DataView(DataView) = DATA_VIEW_DISCRIMINANT,
     Date(Date) = DATE_DISCRIMINANT,
     Error(Error) = ERROR_DISCRIMINANT,
-    FinalizationRegistry(FinalizationRegistryIndex) = FINALIZATION_REGISTRY_DISCRIMINANT,
-    Map(MapIndex) = MAP_DISCRIMINANT,
-    Promise(PromiseIndex) = PROMISE_DISCRIMINANT,
-    Proxy(ProxyIndex) = PROXY_DISCRIMINANT,
-    RegExp(RegExpIndex) = REGEXP_DISCRIMINANT,
-    Set(SetIndex) = SET_DISCRIMINANT,
+    FinalizationRegistry(FinalizationRegistry) = FINALIZATION_REGISTRY_DISCRIMINANT,
+    Map(Map) = MAP_DISCRIMINANT,
+    Promise(Promise) = PROMISE_DISCRIMINANT,
+    Proxy(Proxy) = PROXY_DISCRIMINANT,
+    RegExp(RegExp) = REGEXP_DISCRIMINANT,
+    Set(Set) = SET_DISCRIMINANT,
     SharedArrayBuffer(SharedArrayBuffer) = SHARED_ARRAY_BUFFER_DISCRIMINANT,
-    WeakMap(WeakMapIndex) = WEAK_MAP_DISCRIMINANT,
-    WeakRef(WeakRefIndex) = WEAK_REF_DISCRIMINANT,
-    WeakSet(WeakSetIndex) = WEAK_SET_DISCRIMINANT,
+    WeakMap(WeakMap) = WEAK_MAP_DISCRIMINANT,
+    WeakRef(WeakRef) = WEAK_REF_DISCRIMINANT,
+    WeakSet(WeakSet) = WEAK_SET_DISCRIMINANT,
     Int8Array(TypedArrayIndex) = INT_8_ARRAY_DISCRIMINANT,
     Uint8Array(TypedArrayIndex) = UINT_8_ARRAY_DISCRIMINANT,
     Uint8ClampedArray(TypedArrayIndex) = UINT_8_CLAMPED_ARRAY_DISCRIMINANT,
@@ -108,8 +107,8 @@ pub enum Object {
     AsyncFromSyncIterator = ASYNC_FROM_SYNC_ITERATOR_DISCRIMINANT,
     AsyncIterator = ASYNC_ITERATOR_DISCRIMINANT,
     Iterator = ITERATOR_DISCRIMINANT,
-    Module(ModuleIdentifier) = MODULE_DISCRIMINANT,
-    EmbedderObject(EmbedderObjectIndex) = EMBEDDER_OBJECT_DISCRIMINANT,
+    Module(Module) = MODULE_DISCRIMINANT,
+    EmbedderObject(EmbedderObject) = EMBEDDER_OBJECT_DISCRIMINANT,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
@@ -371,11 +370,11 @@ impl OrdinaryObjectInternalSlots for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_extensible(agent),
+            Object::Map(data) => data.internal_extensible(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_extensible(agent),
+            Object::Set(data) => data.internal_extensible(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -427,11 +426,11 @@ impl OrdinaryObjectInternalSlots for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_set_extensible(agent, value),
+            Object::Map(data) => data.internal_set_extensible(agent, value),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_set_extensible(agent, value),
+            Object::Set(data) => data.internal_set_extensible(agent, value),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -479,11 +478,11 @@ impl OrdinaryObjectInternalSlots for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_prototype(agent),
+            Object::Map(data) => data.internal_prototype(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_prototype(agent),
+            Object::Set(data) => data.internal_prototype(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -537,11 +536,11 @@ impl OrdinaryObjectInternalSlots for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_set_prototype(agent, prototype),
+            Object::Map(data) => data.internal_set_prototype(agent, prototype),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_set_prototype(agent, prototype),
+            Object::Set(data) => data.internal_set_prototype(agent, prototype),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -591,11 +590,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_get_prototype_of(agent),
+            Object::Map(data) => data.internal_get_prototype_of(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_get_prototype_of(agent),
+            Object::Set(data) => data.internal_get_prototype_of(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -653,11 +652,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_set_prototype_of(agent, prototype),
+            Object::Map(data) => data.internal_set_prototype_of(agent, prototype),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_set_prototype_of(agent, prototype),
+            Object::Set(data) => data.internal_set_prototype_of(agent, prototype),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -705,11 +704,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_is_extensible(agent),
+            Object::Map(data) => data.internal_is_extensible(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_is_extensible(agent),
+            Object::Set(data) => data.internal_is_extensible(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -759,11 +758,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_prevent_extensions(agent),
+            Object::Map(data) => data.internal_prevent_extensions(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_prevent_extensions(agent),
+            Object::Set(data) => data.internal_prevent_extensions(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -821,11 +820,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_get_own_property(agent, property_key),
+            Object::Map(data) => data.internal_get_own_property(agent, property_key),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_get_own_property(agent, property_key),
+            Object::Set(data) => data.internal_get_own_property(agent, property_key),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -900,19 +899,15 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_define_own_property(
-                agent,
-                property_key,
-                property_descriptor,
-            ),
+            Object::Map(data) => {
+                data.internal_define_own_property(agent, property_key, property_descriptor)
+            }
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_define_own_property(
-                agent,
-                property_key,
-                property_descriptor,
-            ),
+            Object::Set(data) => {
+                data.internal_define_own_property(agent, property_key, property_descriptor)
+            }
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -966,11 +961,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_has_property(agent, property_key),
+            Object::Map(data) => data.internal_has_property(agent, property_key),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_has_property(agent, property_key),
+            Object::Set(data) => data.internal_has_property(agent, property_key),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -1025,11 +1020,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_get(agent, property_key, receiver),
+            Object::Map(data) => data.internal_get(agent, property_key, receiver),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_get(agent, property_key, receiver),
+            Object::Set(data) => data.internal_get(agent, property_key, receiver),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -1089,11 +1084,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_set(agent, property_key, value, receiver),
+            Object::Map(data) => data.internal_set(agent, property_key, value, receiver),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_set(agent, property_key, value, receiver),
+            Object::Set(data) => data.internal_set(agent, property_key, value, receiver),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -1145,11 +1140,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_delete(agent, property_key),
+            Object::Map(data) => data.internal_delete(agent, property_key),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_delete(agent, property_key),
+            Object::Set(data) => data.internal_delete(agent, property_key),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
@@ -1199,11 +1194,11 @@ impl InternalMethods for Object {
             Object::Arguments => todo!(),
             Object::DataView(_) => todo!(),
             Object::FinalizationRegistry(_) => todo!(),
-            Object::Map(data) => Map::from(data).internal_own_property_keys(agent),
+            Object::Map(data) => data.internal_own_property_keys(agent),
             Object::Promise(_) => todo!(),
             Object::Proxy(_) => todo!(),
             Object::RegExp(_) => todo!(),
-            Object::Set(data) => Set::from(data).internal_own_property_keys(agent),
+            Object::Set(data) => data.internal_own_property_keys(agent),
             Object::SharedArrayBuffer(_) => todo!(),
             Object::WeakMap(_) => todo!(),
             Object::WeakRef(_) => todo!(),
