+++ nova_vm/src/ecmascript/types/language/value.rs
@@ -6,18 +6,17 @@ use crate::{
             to_big_int, to_int32, to_number, to_numeric, to_string, to_uint32,
         },
         builtins::{
-            bound_function::BoundFunction, data_view::DataView, date::Date, error::Error,
-            primitive_objects::PrimitiveObject, shared_array_buffer::SharedArrayBuffer, Array,
-            ArrayBuffer, BuiltinFunction, ECMAScriptFunction,
+            bound_function::BoundFunction, data_view::DataView, date::Date,
+            embedder_object::EmbedderObject, error::Error,
+            finalization_registry::FinalizationRegistry, map::Map, module::Module,
+            primitive_objects::PrimitiveObject, promise::Promise, proxy::Proxy, regexp::RegExp,
+            set::Set, shared_array_buffer::SharedArrayBuffer, weak_map::WeakMap, weak_ref::WeakRef,
+            weak_set::WeakSet, Array, ArrayBuffer, BuiltinFunction, ECMAScriptFunction,
         },
         execution::{Agent, JsResult},
-        scripts_and_modules::module::ModuleIdentifier,
         types::BUILTIN_STRING_MEMORY,
     },
-    heap::indexes::{
-        EmbedderObjectIndex, FinalizationRegistryIndex, MapIndex, PromiseIndex, ProxyIndex,
-        RegExpIndex, SetIndex, TypedArrayIndex, WeakMapIndex, WeakRefIndex, WeakSetIndex,
-    },
+    heap::{indexes::TypedArrayIndex, CompactionLists, HeapMarkAndSweep, WorkQueues},
     SmallInteger, SmallString,
 };
 
@@ -92,16 +91,16 @@ pub enum Value {
     DataView(DataView),
     Date(Date),
     Error(Error),
-    FinalizationRegistry(FinalizationRegistryIndex),
-    Map(MapIndex),
-    Promise(PromiseIndex),
-    Proxy(ProxyIndex),
-    RegExp(RegExpIndex),
-    Set(SetIndex),
+    FinalizationRegistry(FinalizationRegistry),
+    Map(Map),
+    Promise(Promise),
+    Proxy(Proxy),
+    RegExp(RegExp),
+    Set(Set),
     SharedArrayBuffer(SharedArrayBuffer),
-    WeakMap(WeakMapIndex),
-    WeakRef(WeakRefIndex),
-    WeakSet(WeakSetIndex),
+    WeakMap(WeakMap),
+    WeakRef(WeakRef),
+    WeakSet(WeakSet),
 
     // TypedArrays
     Int8Array(TypedArrayIndex),
@@ -123,10 +122,10 @@ pub enum Value {
     Iterator,
 
     // ECMAScript Module
-    Module(ModuleIdentifier),
+    Module(Module),
 
     // Embedder objects
-    EmbedderObject(EmbedderObjectIndex) = 0x7f,
+    EmbedderObject(EmbedderObject) = 0x7f,
 }
 
 /// We want to guarantee that all handles to JS values are register sized. This
@@ -175,8 +174,7 @@ pub(crate) const ECMASCRIPT_FUNCTION_DISCRIMINANT: u8 =
     value_discriminant(Value::ECMAScriptFunction(ECMAScriptFunction::_def()));
 pub(crate) const BOUND_FUNCTION_DISCRIMINANT: u8 =
     value_discriminant(Value::BoundFunction(BoundFunction::_def()));
-pub(crate) const REGEXP_DISCRIMINANT: u8 =
-    value_discriminant(Value::RegExp(RegExpIndex::from_u32_index(0)));
+pub(crate) const REGEXP_DISCRIMINANT: u8 = value_discriminant(Value::RegExp(RegExp::_def()));
 
 pub(crate) const BUILTIN_GENERATOR_FUNCTION_DISCRIMINANT: u8 =
     value_discriminant(Value::BuiltinGeneratorFunction);
@@ -202,23 +200,17 @@ pub(crate) const PRIMITIVE_OBJECT_DISCRIMINANT: u8 =
     value_discriminant(Value::PrimitiveObject(PrimitiveObject::_def()));
 pub(crate) const ARGUMENTS_DISCRIMINANT: u8 = value_discriminant(Value::Arguments);
 pub(crate) const DATA_VIEW_DISCRIMINANT: u8 = value_discriminant(Value::DataView(DataView::_def()));
-pub(crate) const FINALIZATION_REGISTRY_DISCRIMINANT: u8 = value_discriminant(
-    Value::FinalizationRegistry(FinalizationRegistryIndex::from_u32_index(0)),
-);
-pub(crate) const MAP_DISCRIMINANT: u8 = value_discriminant(Value::Map(MapIndex::from_u32_index(0)));
-pub(crate) const PROMISE_DISCRIMINANT: u8 =
-    value_discriminant(Value::Promise(PromiseIndex::from_u32_index(0)));
-pub(crate) const PROXY_DISCRIMINANT: u8 =
-    value_discriminant(Value::Proxy(ProxyIndex::from_u32_index(0)));
-pub(crate) const SET_DISCRIMINANT: u8 = value_discriminant(Value::Set(SetIndex::from_u32_index(0)));
+pub(crate) const FINALIZATION_REGISTRY_DISCRIMINANT: u8 =
+    value_discriminant(Value::FinalizationRegistry(FinalizationRegistry::_def()));
+pub(crate) const MAP_DISCRIMINANT: u8 = value_discriminant(Value::Map(Map::_def()));
+pub(crate) const PROMISE_DISCRIMINANT: u8 = value_discriminant(Value::Promise(Promise::_def()));
+pub(crate) const PROXY_DISCRIMINANT: u8 = value_discriminant(Value::Proxy(Proxy::_def()));
+pub(crate) const SET_DISCRIMINANT: u8 = value_discriminant(Value::Set(Set::_def()));
 pub(crate) const SHARED_ARRAY_BUFFER_DISCRIMINANT: u8 =
     value_discriminant(Value::SharedArrayBuffer(SharedArrayBuffer::_def()));
-pub(crate) const WEAK_MAP_DISCRIMINANT: u8 =
-    value_discriminant(Value::WeakMap(WeakMapIndex::from_u32_index(0)));
-pub(crate) const WEAK_REF_DISCRIMINANT: u8 =
-    value_discriminant(Value::WeakRef(WeakRefIndex::from_u32_index(0)));
-pub(crate) const WEAK_SET_DISCRIMINANT: u8 =
-    value_discriminant(Value::WeakSet(WeakSetIndex::from_u32_index(0)));
+pub(crate) const WEAK_MAP_DISCRIMINANT: u8 = value_discriminant(Value::WeakMap(WeakMap::_def()));
+pub(crate) const WEAK_REF_DISCRIMINANT: u8 = value_discriminant(Value::WeakRef(WeakRef::_def()));
+pub(crate) const WEAK_SET_DISCRIMINANT: u8 = value_discriminant(Value::WeakSet(WeakSet::_def()));
 pub(crate) const INT_8_ARRAY_DISCRIMINANT: u8 =
     value_discriminant(Value::Int8Array(TypedArrayIndex::from_u32_index(0)));
 pub(crate) const UINT_8_ARRAY_DISCRIMINANT: u8 =
@@ -245,11 +237,9 @@ pub(crate) const ASYNC_FROM_SYNC_ITERATOR_DISCRIMINANT: u8 =
     value_discriminant(Value::AsyncFromSyncIterator);
 pub(crate) const ASYNC_ITERATOR_DISCRIMINANT: u8 = value_discriminant(Value::AsyncIterator);
 pub(crate) const ITERATOR_DISCRIMINANT: u8 = value_discriminant(Value::Iterator);
-pub(crate) const MODULE_DISCRIMINANT: u8 =
-    value_discriminant(Value::Module(ModuleIdentifier::from_u32(0)));
-pub(crate) const EMBEDDER_OBJECT_DISCRIMINANT: u8 = value_discriminant(Value::EmbedderObject(
-    EmbedderObjectIndex::from_u32_index(0),
-));
+pub(crate) const MODULE_DISCRIMINANT: u8 = value_discriminant(Value::Module(Module::_def()));
+pub(crate) const EMBEDDER_OBJECT_DISCRIMINANT: u8 =
+    value_discriminant(Value::EmbedderObject(EmbedderObject::_def()));
 
 impl Value {
     pub fn from_str(agent: &mut Agent, str: &str) -> Value {
@@ -510,3 +500,135 @@ impl_value_from_n!(u16);
 impl_value_from_n!(i16);
 impl_value_from_n!(u32);
 impl_value_from_n!(i32);
+
+impl HeapMarkAndSweep for Value {
+    fn mark_values(&self, queues: &mut WorkQueues) {
+        match self {
+            Value::Undefined
+            | Value::Null
+            | Value::Boolean(_)
+            | Value::SmallString(_)
+            | Value::Integer(_)
+            | Value::Float(_)
+            | Value::SmallBigInt(_) => {
+                // Stack values: Nothing to mark
+            }
+            Value::String(idx) => idx.mark_values(queues),
+            Value::Symbol(idx) => idx.mark_values(queues),
+            Value::Number(idx) => idx.mark_values(queues),
+            Value::BigInt(idx) => idx.mark_values(queues),
+            Value::Object(idx) => idx.mark_values(queues),
+            Value::Array(idx) => idx.mark_values(queues),
+            Value::ArrayBuffer(idx) => idx.mark_values(queues),
+            Value::Date(idx) => idx.mark_values(queues),
+            Value::Error(idx) => idx.mark_values(queues),
+            Value::BoundFunction(idx) => idx.mark_values(queues),
+            Value::BuiltinFunction(idx) => idx.mark_values(queues),
+            Value::ECMAScriptFunction(idx) => idx.mark_values(queues),
+            Value::RegExp(idx) => idx.mark_values(queues),
+            Value::PrimitiveObject(idx) => idx.mark_values(queues),
+            Value::Arguments => todo!(),
+            Value::DataView(_) => todo!(),
+            Value::FinalizationRegistry(_) => todo!(),
+            Value::Map(_) => todo!(),
+            Value::Proxy(_) => todo!(),
+            Value::Promise(_) => todo!(),
+            Value::Set(_) => todo!(),
+            Value::SharedArrayBuffer(_) => todo!(),
+            Value::WeakMap(_) => todo!(),
+            Value::WeakRef(_) => todo!(),
+            Value::WeakSet(_) => todo!(),
+            Value::Int8Array(_) => todo!(),
+            Value::Uint8Array(_) => todo!(),
+            Value::Uint8ClampedArray(_) => todo!(),
+            Value::Int16Array(_) => todo!(),
+            Value::Uint16Array(_) => todo!(),
+            Value::Int32Array(_) => todo!(),
+            Value::Uint32Array(_) => todo!(),
+            Value::BigInt64Array(_) => todo!(),
+            Value::BigUint64Array(_) => todo!(),
+            Value::Float32Array(_) => todo!(),
+            Value::Float64Array(_) => todo!(),
+            Value::BuiltinGeneratorFunction => todo!(),
+            Value::BuiltinConstructorFunction => todo!(),
+            Value::BuiltinPromiseResolveFunction => todo!(),
+            Value::BuiltinPromiseRejectFunction => todo!(),
+            Value::BuiltinPromiseCollectorFunction => todo!(),
+            Value::BuiltinProxyRevokerFunction => todo!(),
+            Value::ECMAScriptAsyncFunction => todo!(),
+            Value::ECMAScriptAsyncGeneratorFunction => todo!(),
+            Value::ECMAScriptConstructorFunction => todo!(),
+            Value::ECMAScriptGeneratorFunction => todo!(),
+            Value::AsyncFromSyncIterator => todo!(),
+            Value::AsyncIterator => todo!(),
+            Value::Iterator => todo!(),
+            Value::Module(_) => todo!(),
+            Value::EmbedderObject(_) => todo!(),
+        }
+    }
+
+    fn sweep_values(&mut self, compactions: &CompactionLists) {
+        match self {
+            Value::Undefined
+            | Value::Null
+            | Value::Boolean(_)
+            | Value::SmallString(_)
+            | Value::Integer(_)
+            | Value::Float(_)
+            | Value::SmallBigInt(_) => {
+                // Stack values: Nothing to sweep
+            }
+            Value::String(idx) => idx.sweep_values(compactions),
+            Value::Symbol(idx) => idx.sweep_values(compactions),
+            Value::Number(idx) => idx.sweep_values(compactions),
+            Value::BigInt(idx) => idx.sweep_values(compactions),
+            Value::Object(idx) => idx.sweep_values(compactions),
+            Value::Array(idx) => idx.sweep_values(compactions),
+            Value::ArrayBuffer(idx) => idx.sweep_values(compactions),
+            Value::Date(idx) => idx.sweep_values(compactions),
+            Value::Error(idx) => idx.sweep_values(compactions),
+            Value::BoundFunction(idx) => idx.sweep_values(compactions),
+            Value::BuiltinFunction(idx) => idx.sweep_values(compactions),
+            Value::ECMAScriptFunction(idx) => idx.sweep_values(compactions),
+            Value::RegExp(idx) => idx.sweep_values(compactions),
+            Value::PrimitiveObject(idx) => idx.sweep_values(compactions),
+            Value::Arguments => todo!(),
+            Value::DataView(_) => todo!(),
+            Value::FinalizationRegistry(_) => todo!(),
+            Value::Map(_) => todo!(),
+            Value::Proxy(_) => todo!(),
+            Value::Promise(_) => todo!(),
+            Value::Set(_) => todo!(),
+            Value::SharedArrayBuffer(_) => todo!(),
+            Value::WeakMap(_) => todo!(),
+            Value::WeakRef(_) => todo!(),
+            Value::WeakSet(_) => todo!(),
+            Value::Int8Array(_) => todo!(),
+            Value::Uint8Array(_) => todo!(),
+            Value::Uint8ClampedArray(_) => todo!(),
+            Value::Int16Array(_) => todo!(),
+            Value::Uint16Array(_) => todo!(),
+            Value::Int32Array(_) => todo!(),
+            Value::Uint32Array(_) => todo!(),
+            Value::BigInt64Array(_) => todo!(),
+            Value::BigUint64Array(_) => todo!(),
+            Value::Float32Array(_) => todo!(),
+            Value::Float64Array(_) => todo!(),
+            Value::BuiltinGeneratorFunction => todo!(),
+            Value::BuiltinConstructorFunction => todo!(),
+            Value::BuiltinPromiseResolveFunction => todo!(),
+            Value::BuiltinPromiseRejectFunction => todo!(),
+            Value::BuiltinPromiseCollectorFunction => todo!(),
+            Value::BuiltinProxyRevokerFunction => todo!(),
+            Value::ECMAScriptAsyncFunction => todo!(),
+            Value::ECMAScriptAsyncGeneratorFunction => todo!(),
+            Value::ECMAScriptConstructorFunction => todo!(),
+            Value::ECMAScriptGeneratorFunction => todo!(),
+            Value::AsyncFromSyncIterator => todo!(),
+            Value::AsyncIterator => todo!(),
+            Value::Iterator => todo!(),
+            Value::Module(_) => todo!(),
+            Value::EmbedderObject(_) => todo!(),
+        }
+    }
+}
