+++ nova_vm/src/heap/heap_bits.rs
@@ -2,46 +2,23 @@ use std::borrow::Borrow;
 
 use super::{
     element_array::{ElementArrayKey, ElementsVector},
-    indexes::{
-        ElementIndex, EmbedderObjectIndex, FinalizationRegistryIndex, MapIndex, PromiseIndex,
-        ProxyIndex, RegExpIndex, SetIndex, TypedArrayIndex, WeakMapIndex, WeakRefIndex,
-        WeakSetIndex,
-    },
-    Heap, SymbolHeapData,
+    indexes::{ElementIndex, TypedArrayIndex},
+    Heap,
 };
 use crate::ecmascript::{
     builtins::{
-        bound_function::BoundFunction,
-        data_view::DataView,
-        date::{data::DateHeapData, Date},
-        embedder_object::data::EmbedderObjectHeapData,
-        error::{Error, ErrorHeapData},
-        finalization_registry::data::FinalizationRegistryHeapData,
-        map::{data::MapHeapData, Map},
-        module::{data::ModuleHeapData, Module},
-        primitive_objects::PrimitiveObject,
-        promise::data::PromiseHeapData,
-        proxy::data::ProxyHeapData,
-        regexp::RegExpHeapData,
-        set::{data::SetHeapData, Set},
-        shared_array_buffer::SharedArrayBuffer,
-        typed_array::data::TypedArrayHeapData,
-        weak_map::data::WeakMapHeapData,
-        weak_ref::data::WeakRefHeapData,
-        weak_set::data::WeakSetHeapData,
-        Array, ArrayBuffer, BuiltinFunction, ECMAScriptFunction,
+        bound_function::BoundFunction, data_view::DataView, date::Date,
+        embedder_object::EmbedderObject, error::Error, finalization_registry::FinalizationRegistry,
+        map::Map, module::Module, primitive_objects::PrimitiveObject, promise::Promise,
+        proxy::Proxy, regexp::RegExp, set::Set, shared_array_buffer::SharedArrayBuffer,
+        weak_map::WeakMap, weak_ref::WeakRef, weak_set::WeakSet, Array, ArrayBuffer,
+        BuiltinFunction, ECMAScriptFunction,
     },
     execution::{
-        DeclarativeEnvironment, DeclarativeEnvironmentIndex, EnvironmentIndex, FunctionEnvironment,
-        FunctionEnvironmentIndex, GlobalEnvironment, GlobalEnvironmentIndex, Intrinsics,
-        ObjectEnvironment, ObjectEnvironmentIndex, PrivateEnvironment, PrivateEnvironmentIndex,
-        Realm, RealmIdentifier,
+        DeclarativeEnvironmentIndex, EnvironmentIndex, FunctionEnvironmentIndex,
+        GlobalEnvironmentIndex, ObjectEnvironmentIndex, RealmIdentifier,
     },
-    scripts_and_modules::{
-        module::ModuleIdentifier,
-        script::{Script, ScriptIdentifier},
-        ScriptOrModule,
-    },
+    scripts_and_modules::script::ScriptIdentifier,
     types::{bigint::HeapBigInt, HeapNumber, HeapString, OrdinaryObject, Symbol, Value},
 };
 
@@ -109,30 +86,30 @@ pub(crate) struct WorkQueues {
     pub e_2_6: Vec<(ElementIndex, u32)>,
     pub e_2_8: Vec<(ElementIndex, u32)>,
     pub ecmascript_functions: Vec<ECMAScriptFunction>,
-    pub embedder_objects: Vec<EmbedderObjectIndex>,
+    pub embedder_objects: Vec<EmbedderObject>,
     pub errors: Vec<Error>,
-    pub finalization_registrys: Vec<FinalizationRegistryIndex>,
+    pub finalization_registrys: Vec<FinalizationRegistry>,
     pub function_environments: Vec<FunctionEnvironmentIndex>,
     pub global_environments: Vec<GlobalEnvironmentIndex>,
-    pub maps: Vec<MapIndex>,
-    pub modules: Vec<ModuleIdentifier>,
+    pub maps: Vec<Map>,
+    pub modules: Vec<Module>,
     pub numbers: Vec<HeapNumber>,
     pub object_environments: Vec<ObjectEnvironmentIndex>,
     pub objects: Vec<OrdinaryObject>,
     pub primitive_objects: Vec<PrimitiveObject>,
-    pub promises: Vec<PromiseIndex>,
-    pub proxys: Vec<ProxyIndex>,
+    pub promises: Vec<Promise>,
+    pub proxys: Vec<Proxy>,
     pub realms: Vec<RealmIdentifier>,
-    pub regexps: Vec<RegExpIndex>,
+    pub regexps: Vec<RegExp>,
     pub scripts: Vec<ScriptIdentifier>,
-    pub sets: Vec<SetIndex>,
+    pub sets: Vec<Set>,
     pub shared_array_buffers: Vec<SharedArrayBuffer>,
     pub strings: Vec<HeapString>,
     pub symbols: Vec<Symbol>,
     pub typed_arrays: Vec<TypedArrayIndex>,
-    pub weak_maps: Vec<WeakMapIndex>,
-    pub weak_refs: Vec<WeakRefIndex>,
-    pub weak_sets: Vec<WeakSetIndex>,
+    pub weak_maps: Vec<WeakMap>,
+    pub weak_refs: Vec<WeakRef>,
+    pub weak_sets: Vec<WeakSet>,
 }
 
 impl HeapBits {
@@ -786,768 +763,4 @@ pub(crate) fn sweep_heap_u32_elements_vector_values<const N: usize>(
             false
         }
     });
-}
-
-impl HeapMarkAndSweep for MapIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.maps.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.maps.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for RegExpIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.regexps.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.regexps.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for SetIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.sets.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.sets.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for Value {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        match self {
-            Value::Undefined
-            | Value::Null
-            | Value::Boolean(_)
-            | Value::SmallString(_)
-            | Value::Integer(_)
-            | Value::Float(_)
-            | Value::SmallBigInt(_) => {
-                // Stack values: Nothing to mark
-            }
-            Value::String(idx) => idx.mark_values(queues),
-            Value::Symbol(idx) => idx.mark_values(queues),
-            Value::Number(idx) => idx.mark_values(queues),
-            Value::BigInt(idx) => idx.mark_values(queues),
-            Value::Object(idx) => idx.mark_values(queues),
-            Value::Array(idx) => idx.mark_values(queues),
-            Value::ArrayBuffer(idx) => idx.mark_values(queues),
-            Value::Date(idx) => idx.mark_values(queues),
-            Value::Error(idx) => idx.mark_values(queues),
-            Value::BoundFunction(idx) => idx.mark_values(queues),
-            Value::BuiltinFunction(idx) => idx.mark_values(queues),
-            Value::ECMAScriptFunction(idx) => idx.mark_values(queues),
-            Value::RegExp(idx) => idx.mark_values(queues),
-            Value::PrimitiveObject(idx) => idx.mark_values(queues),
-            Value::Arguments => todo!(),
-            Value::DataView(_) => todo!(),
-            Value::FinalizationRegistry(_) => todo!(),
-            Value::Map(_) => todo!(),
-            Value::Proxy(_) => todo!(),
-            Value::Promise(_) => todo!(),
-            Value::Set(_) => todo!(),
-            Value::SharedArrayBuffer(_) => todo!(),
-            Value::WeakMap(_) => todo!(),
-            Value::WeakRef(_) => todo!(),
-            Value::WeakSet(_) => todo!(),
-            Value::Int8Array(_) => todo!(),
-            Value::Uint8Array(_) => todo!(),
-            Value::Uint8ClampedArray(_) => todo!(),
-            Value::Int16Array(_) => todo!(),
-            Value::Uint16Array(_) => todo!(),
-            Value::Int32Array(_) => todo!(),
-            Value::Uint32Array(_) => todo!(),
-            Value::BigInt64Array(_) => todo!(),
-            Value::BigUint64Array(_) => todo!(),
-            Value::Float32Array(_) => todo!(),
-            Value::Float64Array(_) => todo!(),
-            Value::BuiltinGeneratorFunction => todo!(),
-            Value::BuiltinConstructorFunction => todo!(),
-            Value::BuiltinPromiseResolveFunction => todo!(),
-            Value::BuiltinPromiseRejectFunction => todo!(),
-            Value::BuiltinPromiseCollectorFunction => todo!(),
-            Value::BuiltinProxyRevokerFunction => todo!(),
-            Value::ECMAScriptAsyncFunction => todo!(),
-            Value::ECMAScriptAsyncGeneratorFunction => todo!(),
-            Value::ECMAScriptConstructorFunction => todo!(),
-            Value::ECMAScriptGeneratorFunction => todo!(),
-            Value::AsyncFromSyncIterator => todo!(),
-            Value::AsyncIterator => todo!(),
-            Value::Iterator => todo!(),
-            Value::Module(_) => todo!(),
-            Value::EmbedderObject(_) => todo!(),
-        }
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        match self {
-            Value::Undefined
-            | Value::Null
-            | Value::Boolean(_)
-            | Value::SmallString(_)
-            | Value::Integer(_)
-            | Value::Float(_)
-            | Value::SmallBigInt(_) => {
-                // Stack values: Nothing to sweep
-            }
-            Value::String(idx) => idx.sweep_values(compactions),
-            Value::Symbol(idx) => idx.sweep_values(compactions),
-            Value::Number(idx) => idx.sweep_values(compactions),
-            Value::BigInt(idx) => idx.sweep_values(compactions),
-            Value::Object(idx) => idx.sweep_values(compactions),
-            Value::Array(idx) => idx.sweep_values(compactions),
-            Value::ArrayBuffer(idx) => idx.sweep_values(compactions),
-            Value::Date(idx) => idx.sweep_values(compactions),
-            Value::Error(idx) => idx.sweep_values(compactions),
-            Value::BoundFunction(idx) => idx.sweep_values(compactions),
-            Value::BuiltinFunction(idx) => idx.sweep_values(compactions),
-            Value::ECMAScriptFunction(idx) => idx.sweep_values(compactions),
-            Value::RegExp(idx) => idx.sweep_values(compactions),
-            Value::PrimitiveObject(idx) => idx.sweep_values(compactions),
-            Value::Arguments => todo!(),
-            Value::DataView(_) => todo!(),
-            Value::FinalizationRegistry(_) => todo!(),
-            Value::Map(_) => todo!(),
-            Value::Proxy(_) => todo!(),
-            Value::Promise(_) => todo!(),
-            Value::Set(_) => todo!(),
-            Value::SharedArrayBuffer(_) => todo!(),
-            Value::WeakMap(_) => todo!(),
-            Value::WeakRef(_) => todo!(),
-            Value::WeakSet(_) => todo!(),
-            Value::Int8Array(_) => todo!(),
-            Value::Uint8Array(_) => todo!(),
-            Value::Uint8ClampedArray(_) => todo!(),
-            Value::Int16Array(_) => todo!(),
-            Value::Uint16Array(_) => todo!(),
-            Value::Int32Array(_) => todo!(),
-            Value::Uint32Array(_) => todo!(),
-            Value::BigInt64Array(_) => todo!(),
-            Value::BigUint64Array(_) => todo!(),
-            Value::Float32Array(_) => todo!(),
-            Value::Float64Array(_) => todo!(),
-            Value::BuiltinGeneratorFunction => todo!(),
-            Value::BuiltinConstructorFunction => todo!(),
-            Value::BuiltinPromiseResolveFunction => todo!(),
-            Value::BuiltinPromiseRejectFunction => todo!(),
-            Value::BuiltinPromiseCollectorFunction => todo!(),
-            Value::BuiltinProxyRevokerFunction => todo!(),
-            Value::ECMAScriptAsyncFunction => todo!(),
-            Value::ECMAScriptAsyncGeneratorFunction => todo!(),
-            Value::ECMAScriptConstructorFunction => todo!(),
-            Value::ECMAScriptGeneratorFunction => todo!(),
-            Value::AsyncFromSyncIterator => todo!(),
-            Value::AsyncIterator => todo!(),
-            Value::Iterator => todo!(),
-            Value::Module(_) => todo!(),
-            Value::EmbedderObject(_) => todo!(),
-        }
-    }
-}
-
-impl HeapMarkAndSweep for Map {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.0.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.0.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for Set {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.0.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.0.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ElementsVector {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        match self.cap {
-            ElementArrayKey::Empty => {}
-            ElementArrayKey::E4 => queues.e_2_4.push((self.elements_index, self.len)),
-            ElementArrayKey::E6 => queues.e_2_6.push((self.elements_index, self.len)),
-            ElementArrayKey::E8 => queues.e_2_8.push((self.elements_index, self.len)),
-            ElementArrayKey::E10 => queues.e_2_10.push((self.elements_index, self.len)),
-            ElementArrayKey::E12 => queues.e_2_12.push((self.elements_index, self.len)),
-            ElementArrayKey::E16 => queues.e_2_16.push((self.elements_index, self.len)),
-            ElementArrayKey::E24 => queues.e_2_24.push((self.elements_index, self.len)),
-            ElementArrayKey::E32 => queues.e_2_32.push((self.elements_index, self.len)),
-        }
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.elements_index.into_u32();
-        let shift = match self.cap {
-            ElementArrayKey::Empty => {
-                return;
-            }
-            ElementArrayKey::E4 => compactions.e_2_4.get_shift_for_index(self_index),
-            ElementArrayKey::E6 => compactions.e_2_6.get_shift_for_index(self_index),
-            ElementArrayKey::E8 => compactions.e_2_8.get_shift_for_index(self_index),
-            ElementArrayKey::E10 => compactions.e_2_10.get_shift_for_index(self_index),
-            ElementArrayKey::E12 => compactions.e_2_12.get_shift_for_index(self_index),
-            ElementArrayKey::E16 => compactions.e_2_16.get_shift_for_index(self_index),
-            ElementArrayKey::E24 => compactions.e_2_24.get_shift_for_index(self_index),
-            ElementArrayKey::E32 => compactions.e_2_32.get_shift_for_index(self_index),
-        };
-        self.elements_index = ElementIndex::from_u32(self_index - shift);
-    }
-}
-
-impl HeapMarkAndSweep for DateHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for EmbedderObjectHeapData {
-    fn mark_values(&self, _queues: &mut WorkQueues) {}
-
-    fn sweep_values(&mut self, _compactions: &CompactionLists) {}
-}
-
-impl HeapMarkAndSweep for ErrorHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-        self.message.mark_values(queues);
-        self.cause.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-        self.message.sweep_values(compactions);
-        self.cause.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for FinalizationRegistryHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for MapHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-        self.keys.iter().for_each(|value| value.mark_values(queues));
-        self.values
-            .iter()
-            .for_each(|value| value.mark_values(queues));
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-        self.keys
-            .iter_mut()
-            .for_each(|value| value.sweep_values(compactions));
-        self.values
-            .iter_mut()
-            .for_each(|value| value.sweep_values(compactions));
-    }
-}
-
-impl HeapMarkAndSweep for PromiseHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ProxyHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for RegExpHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for SetHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-        self.set.iter().for_each(|value| value.mark_values(queues));
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-        self.set
-            .iter_mut()
-            .for_each(|value| value.sweep_values(compactions));
-    }
-}
-
-impl HeapMarkAndSweep for SymbolHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.descriptor.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.descriptor.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for TypedArrayHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for WeakMapHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for WeakSetHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for WeakRefHeapData {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.object_index.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.object_index.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for Module {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.0.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.0.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ModuleIdentifier {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.modules.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.modules.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for ModuleHeapData {
-    fn mark_values(&self, _queues: &mut WorkQueues) {}
-
-    fn sweep_values(&mut self, _compactions: &CompactionLists) {}
-}
-
-impl HeapMarkAndSweep for RealmIdentifier {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.realms.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.realms.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for Realm {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.intrinsics().mark_values(queues);
-        self.global_env.mark_values(queues);
-        self.global_object.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.intrinsics_mut().sweep_values(compactions);
-        self.global_env.sweep_values(compactions);
-        self.global_object.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for Intrinsics {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.aggregate_error_prototype().mark_values(queues);
-        self.aggregate_error().mark_values(queues);
-        self.array_prototype_sort().mark_values(queues);
-        self.array_prototype_to_string().mark_values(queues);
-        self.array_prototype_values().mark_values(queues);
-        self.array_prototype().mark_values(queues);
-        self.array().mark_values(queues);
-        self.array_buffer_prototype().mark_values(queues);
-        self.array_buffer().mark_values(queues);
-        self.array_iterator_prototype().mark_values(queues);
-        self.async_from_sync_iterator_prototype()
-            .mark_values(queues);
-        self.async_function_prototype().mark_values(queues);
-        self.async_function().mark_values(queues);
-        self.async_generator_function_prototype()
-            .mark_values(queues);
-        self.async_generator_function().mark_values(queues);
-        self.async_generator_prototype().mark_values(queues);
-        self.async_iterator_prototype().mark_values(queues);
-        self.atomics().mark_values(queues);
-        self.big_int_prototype().mark_values(queues);
-        self.big_int().mark_values(queues);
-        self.big_int64_array().mark_values(queues);
-        self.big_int64_array_prototype().mark_values(queues);
-        self.big_uint64_array().mark_values(queues);
-        self.big_uint64_array_prototype().mark_values(queues);
-        self.boolean_prototype().mark_values(queues);
-        self.boolean().mark_values(queues);
-        self.data_view_prototype().mark_values(queues);
-        self.data_view().mark_values(queues);
-        self.date_prototype_to_utcstring().mark_values(queues);
-        self.date_prototype().mark_values(queues);
-        self.date().mark_values(queues);
-        self.decode_uri().mark_values(queues);
-        self.decode_uricomponent().mark_values(queues);
-        self.encode_uri().mark_values(queues);
-        self.encode_uri_component().mark_values(queues);
-        self.error_prototype().mark_values(queues);
-        self.error().mark_values(queues);
-        self.escape().mark_values(queues);
-        self.eval().mark_values(queues);
-        self.eval_error_prototype().mark_values(queues);
-        self.eval_error().mark_values(queues);
-        self.finalization_registry_prototype().mark_values(queues);
-        self.finalization_registry().mark_values(queues);
-        self.float32_array().mark_values(queues);
-        self.float32_array_prototype().mark_values(queues);
-        self.float64_array().mark_values(queues);
-        self.float64_array_prototype().mark_values(queues);
-        self.function_prototype().mark_values(queues);
-        self.function().mark_values(queues);
-        self.generator_function_prototype_prototype_next()
-            .mark_values(queues);
-        self.generator_function_prototype().mark_values(queues);
-        self.generator_function().mark_values(queues);
-        self.generator_prototype().mark_values(queues);
-        self.int16_array().mark_values(queues);
-        self.int16_array_prototype().mark_values(queues);
-        self.int32_array().mark_values(queues);
-        self.int32_array_prototype().mark_values(queues);
-        self.int8_array().mark_values(queues);
-        self.int8_array_prototype().mark_values(queues);
-        self.is_finite().mark_values(queues);
-        self.is_nan().mark_values(queues);
-        self.iterator_prototype().mark_values(queues);
-        self.json().mark_values(queues);
-        self.map_prototype_entries().mark_values(queues);
-        self.map_prototype().mark_values(queues);
-        self.map().mark_values(queues);
-        self.map_iterator_prototype().mark_values(queues);
-        self.math().mark_values(queues);
-        self.number_prototype().mark_values(queues);
-        self.number().mark_values(queues);
-        self.object_prototype_to_string().mark_values(queues);
-        self.object_prototype().mark_values(queues);
-        self.object().mark_values(queues);
-        self.parse_float().mark_values(queues);
-        self.parse_int().mark_values(queues);
-        self.promise_prototype().mark_values(queues);
-        self.promise().mark_values(queues);
-        self.proxy().mark_values(queues);
-        self.range_error_prototype().mark_values(queues);
-        self.range_error().mark_values(queues);
-        self.reference_error_prototype().mark_values(queues);
-        self.reference_error().mark_values(queues);
-        self.reflect().mark_values(queues);
-        self.reg_exp_prototype_exec().mark_values(queues);
-        self.reg_exp_prototype().mark_values(queues);
-        self.reg_exp().mark_values(queues);
-        self.reg_exp_string_iterator_prototype().mark_values(queues);
-        self.set_prototype_values().mark_values(queues);
-        self.set_prototype().mark_values(queues);
-        self.set().mark_values(queues);
-        self.set_iterator_prototype().mark_values(queues);
-        self.shared_array_buffer_prototype().mark_values(queues);
-        self.shared_array_buffer().mark_values(queues);
-        self.string_prototype_trim_end().mark_values(queues);
-        self.string_prototype_trim_start().mark_values(queues);
-        self.string_prototype().mark_values(queues);
-        self.string().mark_values(queues);
-        self.string_iterator_prototype().mark_values(queues);
-        self.symbol_prototype().mark_values(queues);
-        self.symbol().mark_values(queues);
-        self.syntax_error_prototype().mark_values(queues);
-        self.syntax_error().mark_values(queues);
-        self.throw_type_error().mark_values(queues);
-        self.typed_array_prototype_values().mark_values(queues);
-        self.typed_array_prototype().mark_values(queues);
-        self.typed_array().mark_values(queues);
-        self.typed_array_prototype().mark_values(queues);
-        self.type_error_prototype().mark_values(queues);
-        self.type_error().mark_values(queues);
-        self.type_error_prototype().mark_values(queues);
-        self.uint16_array().mark_values(queues);
-        self.uint16_array_prototype().mark_values(queues);
-        self.uint32_array().mark_values(queues);
-        self.uint32_array_prototype().mark_values(queues);
-        self.uint8_array().mark_values(queues);
-        self.uint8_array_prototype().mark_values(queues);
-        self.uint8_clamped_array().mark_values(queues);
-        self.uint8_clamped_array_prototype().mark_values(queues);
-        self.unescape().mark_values(queues);
-        self.uri_error_prototype().mark_values(queues);
-        self.uri_error().mark_values(queues);
-        self.weak_map_prototype().mark_values(queues);
-        self.weak_map().mark_values(queues);
-        self.weak_ref_prototype().mark_values(queues);
-        self.weak_ref().mark_values(queues);
-        self.weak_set_prototype().mark_values(queues);
-        self.weak_set().mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        OrdinaryObject(self.object_index_base).sweep_values(compactions);
-        BuiltinFunction(self.builtin_function_index_base).sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ScriptIdentifier {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.scripts.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(self_index - compactions.scripts.get_shift_for_index(self_index));
-    }
-}
-
-impl HeapMarkAndSweep for Script {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.realm.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.realm.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ScriptOrModule {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        match self {
-            ScriptOrModule::Script(idx) => idx.mark_values(queues),
-            ScriptOrModule::Module(idx) => idx.mark_values(queues),
-        }
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        match self {
-            ScriptOrModule::Script(idx) => idx.sweep_values(compactions),
-            ScriptOrModule::Module(idx) => idx.sweep_values(compactions),
-        }
-    }
-}
-
-impl HeapMarkAndSweep for DeclarativeEnvironmentIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.declarative_environments.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(
-            self_index
-                - compactions
-                    .declarative_environments
-                    .get_shift_for_index(self_index),
-        );
-    }
-}
-
-impl HeapMarkAndSweep for FunctionEnvironmentIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.function_environments.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(
-            self_index
-                - compactions
-                    .function_environments
-                    .get_shift_for_index(self_index),
-        );
-    }
-}
-
-impl HeapMarkAndSweep for GlobalEnvironmentIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.global_environments.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(
-            self_index
-                - compactions
-                    .global_environments
-                    .get_shift_for_index(self_index),
-        );
-    }
-}
-
-impl HeapMarkAndSweep for ObjectEnvironmentIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        queues.object_environments.push(*self);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        let self_index = self.into_u32();
-        *self = Self::from_u32(
-            self_index
-                - compactions
-                    .object_environments
-                    .get_shift_for_index(self_index),
-        );
-    }
-}
-
-impl HeapMarkAndSweep for PrivateEnvironmentIndex {
-    fn mark_values(&self, _queues: &mut WorkQueues) {
-        todo!()
-    }
-
-    fn sweep_values(&mut self, _compactions: &CompactionLists) {
-        todo!()
-    }
-}
-
-impl HeapMarkAndSweep for EnvironmentIndex {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        match self {
-            EnvironmentIndex::Declarative(idx) => idx.mark_values(queues),
-            EnvironmentIndex::Function(idx) => idx.mark_values(queues),
-            EnvironmentIndex::Global(idx) => idx.mark_values(queues),
-            EnvironmentIndex::Object(idx) => idx.mark_values(queues),
-        }
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        match self {
-            EnvironmentIndex::Declarative(idx) => idx.sweep_values(compactions),
-            EnvironmentIndex::Function(idx) => idx.sweep_values(compactions),
-            EnvironmentIndex::Global(idx) => idx.sweep_values(compactions),
-            EnvironmentIndex::Object(idx) => idx.sweep_values(compactions),
-        }
-    }
-}
-
-impl HeapMarkAndSweep for DeclarativeEnvironment {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.outer_env.mark_values(queues);
-        for binding in self.bindings.values() {
-            binding.value.mark_values(queues);
-        }
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.outer_env.sweep_values(compactions);
-        for binding in self.bindings.values_mut() {
-            binding.value.sweep_values(compactions);
-        }
-    }
-}
-
-impl HeapMarkAndSweep for FunctionEnvironment {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.declarative_environment.mark_values(queues);
-        self.function_object.mark_values(queues);
-        self.new_target.mark_values(queues);
-        self.this_value.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.declarative_environment.sweep_values(compactions);
-        self.function_object.sweep_values(compactions);
-        self.new_target.sweep_values(compactions);
-        self.this_value.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for GlobalEnvironment {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.declarative_record.mark_values(queues);
-        self.global_this_value.mark_values(queues);
-        self.object_record.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.declarative_record.sweep_values(compactions);
-        self.global_this_value.sweep_values(compactions);
-        self.object_record.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for ObjectEnvironment {
-    fn mark_values(&self, queues: &mut WorkQueues) {
-        self.outer_env.mark_values(queues);
-        self.binding_object.mark_values(queues);
-    }
-
-    fn sweep_values(&mut self, compactions: &CompactionLists) {
-        self.outer_env.sweep_values(compactions);
-        self.binding_object.sweep_values(compactions);
-    }
-}
-
-impl HeapMarkAndSweep for PrivateEnvironment {
-    fn mark_values(&self, _queues: &mut WorkQueues) {
-        todo!()
-    }
-
-    fn sweep_values(&mut self, _compactions: &CompactionLists) {
-        todo!()
-    }
 }
